Update EntsoeEnergyPrice.ipynb
Creation of the update_dataset() function: A function was created to update the dataset with new data. This function takes a list of dates, a date format, and the combined DataFrame as parameters, and returns the updated DataFrame.

Use of pd.to_datetime(): Instead of using datetime.strptime() to convert dates, we used pd.to_datetime() for more efficient and consistent date conversion with pandas.

Simplification of date list generation: Instead of using explicit loops, we used list comprehension to generate date lists more concisely.

Removal of unnecessary variables: We removed the df_tempE and df_tempL variables as they are no longer needed with the introduction of the update_dataset() function.

Addition of an "if" statement for a date limit: We added an "if" statement to check if the minimum date in the dataset is greater than or equal to a specific date limit (in this case, January 1, 2012). This ensures that only dates beyond the limit are considered for updating the dataset.

Correction of column mapping: We fixed the line where column mapping is done by removing the assignment of the rename method's return value to the df_comb variable, as the inplace=True option already performs the renaming directly on the DataFrame